issue of strange line-wrap found?
I believe the issue lies in the unicode chars for displaying the exitstatus. No matter the color, the bug happens reproducibly when using the unicode chars on a remote Ubuntu server via ssh. If I use anything different for $_CROSS and $_CHECK all seems to be fine.
This placement of \[ and \] _seems_ to fix the issue too, still using the unicode chars.